fix handling of format merging
multiple formats getting merged into a single format (usually
BSS/VGC) is lossy but a necessary evil to not break any APIs. after
this commit, when multiple formats have the same key we will take
the "highest priority" (usually should == most up to date) data for
the main body of the analysis and then include any info from
outdated formats when they won't override the higher priority values

the theory here is that if an outdated analysis set has the same
name as the newer analysis set the newer set is probably an update
of the old one. also, for good measure, `outdated: true` gets added
to the analysis description of old sets to make it easier to filter
(possibly should get added to the set json as well in case people
dont join them?)

pros to this approach:
- minimally invasive change to existing consumers of the API
- should be soley additive (adding information to sets we already
  have)

cons:
- we lose any overview/comments/credits that were on the merged set
- we lose old sets which share the same name as newer sets (probably
  fine)
- we potentially have close to duplicate sets in certain cases when
  the name changed in a newer version but the set stayed the same
  (can possibly in the future add set deduping to handle this)

Ultimately this is kind of messy but short of rethinking how formats
are handled upstream in PS/Smogon this solution is probably the best
we can do atm

closes #8